fix: Fix and refine existing files

+++ src/Modules/Tours/Explorer.Tours.API/Public/ITourRatingReactionService.cs
@@ -6,7 +6,7 @@ namespace Explorer.Tours.API.Public
     public interface ITourRatingReactionService
     {
         PagedResult<TourRatingReactionDto> GetPaged(int page, int pageSize);
-        PagedResult<TourRatingReactionDto> GetPagedByTourRating(int tourRatingId, int page, int pageSize);
+        PagedResult<TourRatingReactionDto> GetPagedByTourRating(long tourRatingId, int page, int pageSize);
         TourRatingReactionDto Create(TourRatingReactionDto entity);
         TourRatingReactionDto Update(TourRatingReactionDto entity);
         void Delete(long id);
