+++ src/Modules/Tours/Explorer.Tours.Core/Domain/RepositoryInterfaces/ITourRatingReactionRepository.cs
@@ -6,7 +6,7 @@ namespace Explorer.Tours.Core.Domain.RepositoryInterfaces
     public interface ITourRatingReactionRepository
     {
         PagedResult<TourRatingReaction> GetPaged(int page, int pageSize);
-        PagedResult<TourRatingReaction> GetPagedByTourRating(int tourRatingId, int page, int pageSize);
+        PagedResult<TourRatingReaction> GetPagedByTourRating(long tourRatingId, int page, int pageSize);
         TourRatingReaction Get(long id);
         TourRatingReaction Create(TourRatingReaction entity);
         TourRatingReaction Update(TourRatingReaction entity);
